+++ index.js
@@ -16,11 +16,15 @@ module.exports = {
     app.import(app.bowerDirectory + '/matchMedia/matchMedia.js');
     app.import('vendor/propagating.js');
   },
+  contentFor: function(type) {
+    if (type === 'head') {
+      return "<div id='paper-wormhole'></div>";
+    }
+  },
   postprocessTree: function(type, tree) {
     if (type === 'all' || type === 'styles') {
       tree = autoprefixer(tree, { browsers: ['last 2 versions'] });
     }
-
     return tree;
   }
 };